requirements.txt
# SPDX-FileCopyrightText: © 2025 Peter Lemenkov
#
# SPDX-License-Identifier: MIT
python-telegram-bot>=20.0
